requirements.txt
# Base packages
langchain
langchain-community
langchain-openai
openai
pydantic==1.10.15
fastapi==0.111.1
uvicorn==0.30.5

# AWS
boto3==1.26.10
botocore==1.29.10

# PDF parsing and web scraping
pypdf==4.2.0 
beautifulsoup4==4.12.3
requests==2.32.3
validators==0.33.0

# Vector store
faiss-cpu==1.8.0

# Utils
python-dotenv==1.0.0
pyyaml==6.0.2

# Metrics and Observability
prometheus-client==0.19.0

# Maybe needed
# chromadb==0.4.22
# unstructured==0.18.31
